pyproject.toml
[tool.poetry]
name = "python-rag-emma"
version = "0.1.0"
description = "This is a RAG implementation example using OpenAI's ChatGPT of answer questions of varying complexity about the Jane Austen book Emma."
authors = ["jezlane <jeremy.rs.lane@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
python-dotenv = "^1.0.1"
langchain = "^0.1.20"
openai = "^1.30.1"
chromadb = "^0.5.0"
tiktoken = "^0.7.0"
langchain-openai = "^0.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
